requirements.txt
tensorflow==2.15.0
keras==2.15.0
flask==3.0.3
numpy==1.26.4
opencv-python==4.9.0.80
Pillow==10.3.0
scikit-learn==1.4.2
matplotlib==3.8.4
kaggle==1.6.14
tqdm==4.66.4
scipy==1.13.0
